+++ analysis/__init__.py
@@ -1,1 +1,2 @@
+from . import _get_data
 from . import _utils
