+++ analysis/_utils.py
@@ -1,49 +1,10 @@
-import json
-
 import streamlit as st
-import pandas as pd
-import requests
 
 
-@st.cache_resource
-def get_regattas():
-    return pd.json_normalize(
-        json.loads(
-            requests.get("http://www.sapsailing.com/sailingserver/api/v1/regattas").text
-        )
-    )
-
-
-@st.cache_resource
-def get_races(regatta):
-    data = json.loads(
-        requests.get(
-            f"http://www.sapsailing.com/sailingserver/api/v1/regattas/{regatta}/races"
-        ).text
-    )
-
-    if len(data["races"]) > 0:
-        return [race["name"].strip() for race in data["races"]]
-    else:
-        return [regatta]
-
-
-@st.cache_resource
-def get_start_data(regatta, race):
-    try:
-        return json.loads(
-            requests.get(
-                f"https://www.sapsailing.com/sailingserver/api/v1/regattas/{regatta}/races/{race}/startanalysis"
-            ).text
-        )
-    except json.JSONDecodeError:
-        return None
-
-
-@st.cache_resource
-def get_mark_data(regatta, race):
-    return json.loads(
-        requests.get(
-            f"https://www.sapsailing.com/sailingserver/api/v1/regattas/{regatta}/races/{race}/markpassings"
-        ).text
-    )
+def hide_streamlit():
+    hide_streamlit_style = """
+            <style>
+            footer {visibility: hidden;}
+            </style>
+            """
+    st.markdown(hide_streamlit_style, unsafe_allow_html=True)
